Shots: added feature to "waterfall" shots
Each shot maintains a table of all the modes that set it to enable or
disable, the profile, etc. If a shot is hit in one mode, it will
waterfall that hit down to the other modes. This can be blocked at any
level with a ‘block’ setting in the shot profile.