Enhancement to stale slide removal
Now slides with expire times will not be counted in the “one slide per
mode” calculation, meaning MPF will leave other slides alone if the
active slide has a removal timer. This will prevent slides from being
removed that are not set to persistent if they’re only temporarily
replaced with a slide with an expire time.