RoundTrip Yaml Additions + AUTHORS Update
Added the modifications to support round trip YAML writing for the
mpf-wizard.  Including:
-Added the optional parameter "round_trip" to the load methods, it is
False by default.  If set to true, the RoundTripLoader will be used
instead of the Loader.
- Added type detection for the ruamel custom classes to the
string_to_list IMPLEMENTATION, added as a string comparison so the
ruamel classes didnt have to directly be imported
- Added a static method called save_to_str to the YamlInterface class to
support writing back out the YAML file.
- Added a unit test for the new code I added, testing the load and save
of a yaml string
- Added myself to the AUTHORS file
- Removed a flagrant abuse of commenting from the mpfconfig.yaml file ;)