Created standalone config_players to be used for shows and config files
Simplified shows so all the logic for each thing they play in each step
is in the associated config_player instead of the show. That way the
same config player is used to play a step in a show as well as an entry
in the config file. Also simplified how lights and leds are updated,
moving the logic to do so to the device and out of the show controller.
Renamed process_config2 to validate_config