Enforce that show_tokens passed to show match what the show wants
I only made this raise an exception if there are invalid tokens passed.
If there are not enough tokens passed, it’s still ok. Reason is that
our default “off” show has tokens for leds, led, light, and lights, so
it can be used as a generic “catch off” show to turn anything off. If
we want to change that, we’ll have to give this some more thought since
lots of people use that. #216