timer: restore using player var
turns out a lot of people were using the player var for timer text on
slides which is elegant if you have a slide that was not created from
the tick event. So now the timer will also set a player var if a player
exists, but it doesn’t change how it works either way.