Added ability to have autofires self timeout.
It should now be possible to have autofire devices automatically disable themselves for a configurable amount of seconds if they are activated a configurable amount of times in a configurable amount of seconds. Basically if activated X times in Y seconds, disable for Z seconds. X, Y, and Z are options in the config file.