Moved import, changed Timer to delay, and fixed race.
Moved "from time import time as _time" from inside the _initialize method to the top of the file so Pylint is happy. Removed Timer from threading and added delay from DelayManger in its place so it doesn't interfere with asyncio. Added delay removal to disable method so as to prevent race.