Update multiball_lock.py
I shortened the function to only test for virtual_only and min_virtual_physical and removed any part of the function that took away balls from the lock. The only problem I am running into is that MYPY is declared false so BallDevice is not imported that means that the function request_ball cannot be called. I added an import statement above it so that there were no errors.