Fixed issue with setting number of LEDs per node
The comment says to set LEDs to 96 and switches to 60.  This wasn't what was happening in the code. As a result, some of the LEDs on some nodes weren't working if their indices were too large. Should mpf auto-detect the number of LEDs/switches based on what's in the config file and set this accordingly?