Fixed error related to show pools
Set the param when def show pools, and returning show pools to include
start_running=True, and start_running value returned.  This allows the
show pool to run, as it previously threw an error that the param was
missing.  Now if the show_player sets start_running to false, it
behaves.  But defaults, like show_player, to start running.