Timeout feature for logic blocks
This includes a new timeout feature for logic blocks:
logic_block_timeout.  If a certain amount of time elapses since the
logic block was enabled, it will reset to its default state, and restart
the timer.  For advanced changes between steps, that will need to be
done with additional logic, or future enhancement to this feature. Test
includes an example for all three types of logic blocks.