Added Machine Vars for Date and Time
Added machine vars for year, month, day, hour, minute, and second.
New async function was created to update this time every second, so when
you call it in MPF, it will always know the current time when the action
is taken.  Tests were updated to ensure that the machine var time
matches the actual time of the machine.

In the future, we will need to give a setting to write to the system
clock.  But for now, the user would need to set the system clock though
the OS/System, and MPF will read that value into a machine var.