Added hurry up and grace period features to the multiball ball save. This allows the user to run shows off the hurry up, as well as provide a grace period for saves.
The tests for test_Modes are failing when I run bulk, but run fine when I run just them.  So not sure what is going on there.  If you also see this issue, please let me know what I changed or need to do to fix that issue.

If this is approved and merged, I would also like to go add similar features for add_a_ball to provide a different and shorter ball save feature.