Added blinkenlights: and blinkenlight_player:
The purpose of blinkenlights are to allow multiple  modes to be able to control the same RGB light.
In this way, if multiple modes share the same shot, those modes can each add their own color to the
light for that shot, and the shot will flash all the different colors to let the player know that
the same shot will be scored for multiple modes. (For example, a multiball and a mode sharing the
same ramp shot.)
Modes can add colors to a blinkenlight with the blinkenlight_player, and the blinkenlight will
cycle through all the colors that have been added.  Modes can also remove colors, or remove all
colors, effectively turning the blinkenlight off.