Adds Multiball Ball Save for Add A Ball
This adds a unique ball save feature for add a ball.  This uses the same
event calls, so you can trigger your shows off the same events for
normal ball save and add a ball ball save.  The add a ball ball save
will not run (or add time) if the ball save is already running.  But if
the normal ball save has expired, it will start this ball save at the
add a ball configured times for shoot again, hurry up, and grace period.