Adds a default show bl_color
The bl_color show takes a token for the light (led, light, blinkenlight,
etc.) as well as the color that is being passed.  It calls to the
blinkenlights player only, but can take any of the standard light
tokens.  It also includes some error handling for blinkenlight_player to
verify its a type, and not one of the blank strings when adding or
removing.