Checks if events exist before setting to 0
The _load_defaults was setting all events that were configured to 0 when
this was called (which happens on every _initiatlize.  This changes it
to be the same logic as switches and players.  It checks if the event is
already in the list.  If it is not, it creates it and sets it to 0.  If
it is in the list, it is ignored.  This also works for the reset because
it blanks the entire list out first, then recreates.  So when it does
this, all values would be missing, and thus are created with a value of
0.