Update match.py to keep winner_number
added a single line to temporarily create a machine variable 'match_number' so the winner number would be accessible during attract mode after match mode ended for the purpose of keeping the match light lit on an em.  I tried to figure out how to have it store this in machine_vars.yaml and reload it next time you start the machine, but that is way way over my head.  not sure if info_lights could work with this as well, but we would only want it lit during attract mode.