Syncs up chimes with reel coil pulse
Adds conssitency for the will_xx _xxing and _xxed format to the
advancement of the score reel.  Changed the event handler for chime
pulses to be the _advancing event, which triggers at the same time as
the coil advance pulse.

The only potential break is if someone is queueing custom events off the
reel_(name)_advance event as this is now _advanceD (past tense).