Save High Score Variables in high_scores.yaml
This extends the high score mode to allow the user to define vars that they want to be saved along with the initials and title of the high score that is achieved.  The user can define any number of player or machine vars that should be saved along with the files.  These are written into the high_scores.yaml file as a third tuple.  The code parses the list and for those with two elements (no vars) or three elements (with vars) it will create the machine variables so that it can be used on attract mode, etc.  Then it will save to the file and update the machine vars after the game ends.